Admin UI Revamp + Card binding fix
Users could bind new maps on the webui but those new cards couldn't be used to play on the cab itself (due to how the card id is converted by the reader). As a fix, binding a valid eamusement card now also binds a second cardid in the reader's format.